requirements.txt
streamlit
pandas
numpy
scikit-learn
xgboost
prophet
matplotlib
openai
python-dotenv